pyproject.toml
[tool.poetry]
name = "fastapi-application"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

bcrypt = "^4.0.1"
python-jose = "^3.3.0"
pydantic= "1.10.11"
fastapi = "0.99.1"
uvicorn= {extras = ["standard"], version = "^0.20.0"}
python-multipart = "^0.0.6"
pymongo = {extras = ["srv"], version = "^4.3.3"}
orjson = "^3.9.1"
langchain = "0.0.206"
openai = "^0.27.8"
chromadb = "0.3.26"
streamlit = "^1.24.1"
google-cloud-translate = "^3.11.2"
azure-cognitiveservices-speech = "^1.30.0"
requests = "^2.31.0"
google-cloud-storage = "^2.10.0"
sdk = "^1.0.0"
httpx = "^0.26.0"
pypdf2 = "^3.0.1"


[tool.poetry.group.dev.dependencies]
types-python-jose = "^3.3.4.7"
ruff = "^0.0.271"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
preinstall = "export OPENBLAS_CORETYPE=ARMV8"
